+++ src/lib.rs
@@ -2,7 +2,7 @@
 #![doc = r#"
 # `wasm-tracing`
 
-Leverages tracing to proilfe wasm performance via `console`.
+Leverages tracing to profile wasm performance via `console`.
 
 ## Usage
 
@@ -25,6 +25,13 @@ pub fn start() -> Result<(), JsValue> {
     Ok(())
 }
 ```
+
+## Features
+
+| Feature | Function |
+| ------------ | -------- |
+| `rayon` | Logs the rayon worker thread's index within its current pool. |
+| `tracing-log` | Provides complete Metadata via tracing-log's [`NormalizeEvent::normalized_metadata`](tracing_log::NormalizeEvent) method. |
 "#]
 
 use tracing::dispatcher::SetGlobalDefaultError;
@@ -43,10 +50,7 @@ pub use layer::*;
 pub(crate) mod recorder;
 /// Re-exports of common types
 pub mod prelude {
-    pub use super::{
-        config::{ConsoleConfig, WasmLayerConfig},
-        layer::WasmLayer,
-    };
+    pub use super::{config::WasmLayerConfig, layer::WasmLayer};
 }
 
 #[wasm_bindgen]
@@ -57,10 +61,6 @@ extern "C" {
     fn measure(name: String, startMark: String) -> Result<(), JsValue>;
     #[wasm_bindgen(js_namespace = console, js_name = log)]
     fn log1(message: String);
-    #[wasm_bindgen(js_namespace = console, js_name = log)]
-    fn log2(message1: &str, message2: &str);
-    #[wasm_bindgen(js_namespace = console, js_name = log)]
-    fn log3(message1: &str, message2: &str, message3: &str);
     #[wasm_bindgen(js_namespace = console, js_name = log)]
     fn log4(message1: String, message2: &str, message3: &str, message4: &str);
     #[wasm_bindgen(js_namespace = console, js_name = debug)]
@@ -106,9 +106,22 @@ fn mark_name(id: &tracing::Id) -> String {
 }
 
 #[doc = r#"
-    Set the global default recorder with [tracing::subscriber::set_global_default]. Panics if the [WasmLayer] cannot be constructed.
+Set the global default recorder with [tracing::subscriber::set_global_default]. Panics if the [WasmLayer] cannot be constructed.
 
-    Panics if a global default is already set.
+Panics if a global default is already set.
+
+## NOTE
+
+It is discouraged by `tracing` for libraries (such at this one) to call [`tracing::subscriber::set_global_default`].
+This function does so, as it is a convenience. If you are a library, please follow the advice of `tracing`.
+
+If you would like to use multiple layers, use this code:
+```rust
+use tracing_subscriber::layer::*;
+use tracing_subscriber::registry::*;
+use wasm_tracing::prelude::*;
+
+tracing::subscriber::set_global_default(Registry::default().with(WasmLayer::default())).unwrap();
 "#]
 pub fn set_as_global_default() {
     tracing::subscriber::set_global_default(
@@ -136,6 +149,19 @@ pub fn start() -> Result<(), JsValue> {
     Ok(())
 }
 ```
+
+## NOTE
+
+It is discouraged by `tracing` for libraries (such at this one) to call [`tracing::subscriber::set_global_default`].
+This function does so, as it is a convenience. If you are a library, please follow the advice of `tracing`.
+
+If you would like to use multiple layers, use this code:
+```rust
+use tracing_subscriber::layer::*;
+use tracing_subscriber::registry::*;
+use wasm_tracing::prelude::*;
+
+tracing::subscriber::set_global_default(Registry::default().with(WasmLayer::default())).unwrap();
 "#]
 pub fn try_set_as_global_default() -> Result<(), SetGlobalDefaultError> {
     tracing::subscriber::set_global_default(
@@ -145,6 +171,7 @@ pub fn try_set_as_global_default() -> Result<(), SetGlobalDefaultError> {
 
 #[doc = r#"
 Given a [`WasmLayerConfig`], set WASM to be the default layer for a [Registry].
+
 
 ## Example
 
@@ -158,12 +185,26 @@ use tracing::Level;
 pub fn start() -> Result<(), JsValue> {
     console_error_panic_hook::set_once();
 
-    let config = WasmLayerConfig::new().set_report_logs_in_timings(true).set_max_level(Level::ERROR).to_owned();
+    let config = WasmLayerConfig::new().remove_timings().with_max_level(Level::ERROR);
 
     let _ = wasm_tracing::set_as_global_default_with_config(config);
 
     Ok(())
 }
+```
+
+## NOTE
+
+It is discouraged by `tracing` for libraries (such at this one) to call [`tracing::subscriber::set_global_default`].
+This function does so, as it is a convenience. If you are a library, please follow the advice of `tracing`.
+
+If you would like to use multiple layers, use this code:
+```rust
+use tracing_subscriber::layer::*;
+use tracing_subscriber::registry::*;
+use wasm_tracing::prelude::*;
+
+tracing::subscriber::set_global_default(Registry::default().with(WasmLayer::default())).unwrap();
 ```
 "#]
 pub fn set_as_global_default_with_config(